+++ packages/web3-eth/test/unit/rpc_method_wrappers/get_chain_id.test.ts
@@ -16,12 +16,12 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { Web3Context } from 'web3-core';
 import { DEFAULT_RETURN_FORMAT, FMT_BYTES, FMT_NUMBER, format } from 'web3-utils';
+import { Web3EthExecutionAPI } from 'web3-types';
+import { ethRpcMethods } from 'web3-rpc-methods';
 
-import { getChainId as rpcMethodsGetChainId } from '../../../src/rpc_methods';
-import { Web3EthExecutionAPI } from '../../../src/web3_eth_execution_api';
 import { getChainId } from '../../../src/rpc_method_wrappers';
 
-jest.mock('../../../src/rpc_methods');
+jest.mock('web3-rpc-methods');
 
 describe('getChainId', () => {
 	let web3Context: Web3Context<Web3EthExecutionAPI>;
@@ -32,7 +32,7 @@ describe('getChainId', () => {
 
 	it('should call rpcMethods.getBlockNumber with expected parameters', async () => {
 		await getChainId(web3Context, DEFAULT_RETURN_FORMAT);
-		expect(rpcMethodsGetChainId).toHaveBeenCalledWith(web3Context.requestManager);
+		expect(ethRpcMethods.getChainId).toHaveBeenCalledWith(web3Context.requestManager);
 	});
 
 	it('should format mockRpcResponse using provided return format', async () => {
@@ -43,7 +43,7 @@ describe('getChainId', () => {
 			mockRpcResponse,
 			expectedReturnFormat,
 		);
-		(rpcMethodsGetChainId as jest.Mock).mockResolvedValueOnce(mockRpcResponse);
+		(ethRpcMethods.getChainId as jest.Mock).mockResolvedValueOnce(mockRpcResponse);
 
 		const result = await getChainId(web3Context, expectedReturnFormat);
 		expect(result).toBe(expectedFormattedResult);