+++ packages/web3-net/CHANGELOG.md
@@ -41,6 +41,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   `web3-rpc-methods` dependency (#5441)
 
+### Changed
+
+-   `Web3NetAPI` is now imported from `web3-types` instead of `web3_net_api.ts` (#5441)
+-   Replace the imported methods from `rpc_methods.ts` with `netRpcMethods` imports from `web3-rpc-methods` (#5441)
+
 ### Removed
 
 -   `rpcMethods` export, these methods are now exported via `web3-rpc-methods` as `netRpcMethods` (#5441)